- use fourier for continuous conditional and embedding for discrete conditional - fix attention norm and rope - change some stuff

+++ src/osu_fusion/modules/positional_embeddings.py
@@ -19,7 +19,7 @@ class SinusoidalPositionEmbedding(nn.Module):
         half_dim = self.dim // 2
         emb = math.log(self.theta) / (half_dim - 1)
         emb = torch.exp(torch.arange(half_dim, device=device) * -emb)
-        emb = x[:, None] * emb[None, :]
+        emb = x.float()[:, None] * emb[None, :]
         emb = torch.cat([emb.sin(), emb.cos()], dim=-1)
         return emb
